Change: A ViewController now owns it's view
Previously, the ViewController collection had to manage to load, inject and destroy a ViewController's view. Methods such as ViewControllerStack::pushViewControllerFrom were used to load an external view, instantiate a ViewController based on a class name stored in the view's data-attribute. Theses methods are gone. The ViewController now handles how its view is loaded and when it's destroyed. The ViewController's constructor receives one parameter, which can be an URL to load a remote view or an element to build its view upon.

Also, bindChild* and didBindChild* methods were removed. The job is now carefully done by addChild* making sure the element to add is not already present in the DOM before adding it.